requirements.txt
# FastAPI and web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Database
sqlalchemy>=2.0.36

# LangChain and AI - Latest stable versions
langchain>=0.3.27
langchain-openai>=0.3.33
openai>=1.107.0

# Data validation and serialization
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Rate limiting and middleware
slowapi>=0.1.9
redis>=5.2.0

# Utilities
python-dotenv>=1.0.1
requests>=2.32.0
tenacity>=9.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0

# Development
black>=24.10.0
flake8>=7.1.0
mypy>=1.13.0